% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apptest.R
\name{apptest}
\alias{apptest}
\title{Start app test}
\usage{
apptest(name = "app", js = NA, css = NA, port = 8880)
}
\arguments{
\item{name}{app name}

\item{js}{js file for testing}

\item{css}{css file for test}

\item{port}{port to serve app}
}
\value{
an app environment
}
\description{
Start app test
}
