% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwa.R
\name{create_pwa}
\alias{create_pwa}
\title{Create pwa instance}
\usage{
create_pwa(indexTag = NULL, publishingSite, folder = "docs", icon512 = NULL)
}
\arguments{
\item{indexTag}{index page tag}

\item{publishingSite}{publishing site's url}

\item{folder}{local folder to contain the website}

\item{icon512}{local file path to a 512x512 icon}
}
\value{
an pwa instance with properties $folder, $swjs, $webmanifest, and with methods $index() turning page_index() into an index page tag with pwa setup, $updateManifest(), $updateServiceWorker(), and $updateIcons(icon512).
}
\description{
an instance with methods
}
