% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appservice.R
\name{create_testapp}
\alias{create_testapp}
\title{Create a web app from shiny tag}
\usage{
create_testapp(tag = .Last.value, folderpath = "temp", htmlfile = "index.html")
}
\arguments{
\item{tag}{a shiny tag}

\item{folderpath}{folder path}

\item{htmlfile}{html file name.}
}
\value{
the full path to the app which is file.path(folderpath, htmlfile)
}
\description{
Create a web app from shiny tag
}
