% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{\%>>\%}
\alias{\%>>\%}
\title{An enhance |> operator. When LHS is not available, it will load_all(".") first.}
\usage{
x \%>>\% y
}
\arguments{
\item{x}{LHS}

\item{y}{RHS}
}
\value{
none.
}
\description{
An enhance |> operator. When LHS is not available, it will load_all(".") first.
}
