% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{\%L\%}
\alias{\%L\%}
\title{An enhance |> operator. It will load_all(".") first.}
\usage{
x \%L\% y
}
\arguments{
\item{x}{LHS}

\item{y}{RHS}
}
\value{
none.
}
\description{
An enhance |> operator. It will load_all(".") first.
}
