% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appservice.R
\name{start_appservice}
\alias{start_appservice}
\title{Start to serve web app at temp/ and lauch chromedriver for inspection}
\usage{
start_appservice(
  appPath = "./temp/index.html",
  chromedrivePort = 9515,
  app_port = 4321,
  google = F
)
}
\value{
a job id
}
\description{
Start to serve web app at temp/ and lauch chromedriver for inspection
}
