% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appservice.R
\name{webdriverChromeSession}
\alias{webdriverChromeSession}
\title{chromedriver webdrive session starter}
\usage{
webdriverChromeSession(chromedrivePort = 9515)
}
\value{
an environment with $start_session() to start a webdriver session connected to chromedriver; with $kill_chrome() to kill chromedriver process.
}
\description{
chromedriver webdrive session starter
}
