% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apptest.R
\name{apptest}
\alias{apptest}
\title{Start app test}
\usage{
apptest(name = "app", port = 8880, google = F)
}
\arguments{
\item{name}{app name}

\item{port}{port to serve app}

\item{js}{js file for testing}

\item{css}{css file for test}
}
\value{
an app environment
}
\description{
Start app test
}
