% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appservice.R
\name{create_testapp}
\alias{create_testapp}
\title{Create a web app from shiny tag}
\usage{
create_testapp(appmeta, htmlfile = "temp/index.html")
}
\arguments{
\item{appmeta}{app$meta}

\item{htmlfile}{html file name.}

\item{folderpath}{folder path}
}
\value{
the full path to the app which is file.path(folderpath, htmlfile)
}
\description{
Create a web app from shiny tag
}
